requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
langchain>=0.0.350
langchain-openai>=0.0.2
langchain-community==0.0.16
openai>=1.10.0,<2.0.0
httpx>=0.25.1
sqlalchemy>=2.0.23
aiosqlite==0.19.0
pydantic>=2.5.1
pydantic-settings>=2.1.0
requests>=2.31.0
python-multipart>=0.0.6
typing-extensions>=4.8.0
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
networkx
loguru>=0.7.2 